requirements.txt
fastapi
uvicorn[standard]
python-multipart
librosa
soundfile
onnxruntime
numpy
scipy
jinja2
python-dotenv
pydantic-settings
